Conjugation typeclasses. Orthogonal reqs for Complex/Quaternion.
Added a Conjugation typeclass, with instances for Complex/Quaternion,
and trivial instances for types having IsReal instances.

Made Complex require (and implement) commutative instances. Made the
Complex requirements stricter, as Complex operations now require at
least a CRing.

Made Quaternion require CRing as well.
