+++ core/src/main/scala/spire/math/Complex.scala
@@ -11,16 +11,16 @@ import spire.syntax.order._
 import scala.math.{ScalaNumber, ScalaNumericConversions}
 
 object Complex extends ComplexInstances {
-  def i[@sp(Float, Double) T](implicit T: Rig[T]): Complex[T] =
+  def i[@sp(Float, Double) T](implicit T: CRing[T]): Complex[T] =
     new Complex(T.zero, T.one)
 
-  def one[@sp(Float, Double) T](implicit T: Rig[T]): Complex[T] =
+  def one[@sp(Float, Double) T](implicit T: CRing[T]): Complex[T] =
     new Complex(T.one, T.zero)
 
-  def zero[@sp(Float, Double) T](implicit T: Semiring[T]): Complex[T] =
+  def zero[@sp(Float, Double) T](implicit T: CRing[T]): Complex[T] =
     new Complex(T.zero, T.zero)
 
-  def fromInt[@sp(Float, Double) T](n: Int)(implicit f: Ring[T]): Complex[T] =
+  def fromInt[@sp(Float, Double) T](n: Int)(implicit f: CRing[T]): Complex[T] =
     new Complex(f.fromInt(n), f.zero)
 
   implicit def intToComplex(n: Int): Complex[Double] = new Complex(n.toDouble, 0.0)
@@ -39,8 +39,8 @@ object Complex extends ComplexInstances {
   def polar[@sp(Float, Double) T: Field: Trig](magnitude: T, angle: T): Complex[T] =
     new Complex(magnitude * Trig[T].cos(angle), magnitude * Trig[T].sin(angle))
 
-  def apply[@sp(Float, Double) T: Semiring](real: T): Complex[T] =
-    new Complex(real, Semiring[T].zero)
+  def apply[@sp(Float, Double) T: CRing](real: T): Complex[T] =
+    new Complex(real, CRing[T].zero)
 
   def rootOfUnity[@sp(Float, Double) T](n: Int, x: Int)(implicit f: Field[T], t: Trig[T]): Complex[T] = {
     if (x == 0) return one[T]
@@ -84,6 +84,12 @@ object Complex extends ComplexInstances {
   }
 }
 
+/** Complex numbers. Depending on the underlying scalar T, can represent the Gaussian integers (T = BigInt/SafeLong),
+  * the Gaussian rationals (T = Rational) or the complex number field (T: Field).
+  *
+  * Note that we require T to be at least CRing, a commutative ring, so the implementation below is slightly
+  * less general than the Cayley-Dickson construction.
+  */
 @SerialVersionUID(0L)
 final case class Complex[@sp(Float, Double) T](real: T, imag: T)
     extends ScalaNumber with ScalaNumericConversions with Serializable { lhs =>
@@ -100,12 +106,14 @@ final case class Complex[@sp(Float, Double) T](real: T, imag: T)
 
   def abs(implicit f: Field[T], n: NRoot[T], s: Signed[T]): T = hypot(real, imag)
 
+  def absSquare(implicit r: CRing[T]): T = real*real + imag*imag
+
   def arg(implicit f: Field[T], s: Signed[T], t: Trig[T]): T =
     if (isZero) f.zero else t.atan2(imag, real)
 
   def norm(implicit f: Field[T], n: NRoot[T], s: Signed[T]): T = hypot(real, imag)
 
-  def conjugate(implicit f: Rng[T]): Complex[T] = new Complex(real, -imag)
+  def conjugate(implicit f: CRing[T]): Complex[T] = new Complex(real, -imag)
 
   def asTuple: (T, T) = (real, imag)
   def asPolarTuple(implicit f: Field[T], n: NRoot[T], s: Signed[T], t: Trig[T]): (T, T) = (abs, arg)
@@ -117,11 +125,11 @@ final case class Complex[@sp(Float, Double) T](real: T, imag: T)
   def eqv(b: Complex[T])(implicit o: Eq[T]): Boolean = real === b.real && imag === b.imag
   def neqv(b: Complex[T])(implicit o: Eq[T]): Boolean = real =!= b.real || imag =!= b.imag
 
-  def unary_-(implicit r: Rng[T]): Complex[T] = new Complex(-real, -imag)
+  def unary_-(implicit r: CRing[T]): Complex[T] = new Complex(-real, -imag)
 
-  def +(rhs: T)(implicit r: Semiring[T]): Complex[T] = new Complex(real + rhs, imag)
-  def -(rhs: T)(implicit r: Rng[T]): Complex[T] = new Complex(real - rhs, imag)
-  def *(rhs: T)(implicit r: Semiring[T]): Complex[T] = new Complex(real * rhs, imag * rhs)
+  def +(rhs: T)(implicit r: CRing[T]): Complex[T] = new Complex(real + rhs, imag)
+  def -(rhs: T)(implicit r: CRing[T]): Complex[T] = new Complex(real - rhs, imag)
+  def *(rhs: T)(implicit r: CRing[T]): Complex[T] = new Complex(real * rhs, imag * rhs)
   def /(rhs: T)(implicit r: Field[T]): Complex[T] = new Complex(real / rhs, imag / rhs)
 
     /* TODO: does it make sense? Should match the behavior on Gaussian integers.
@@ -147,13 +155,13 @@ final case class Complex[@sp(Float, Double) T](real: T, imag: T)
       Complex.polar(abs fpow e, arg * e)
     }
 
-  def +(b: Complex[T])(implicit r: Semiring[T]): Complex[T] =
+  def +(b: Complex[T])(implicit r: CRing[T]): Complex[T] =
     new Complex(real + b.real, imag + b.imag)
 
-  def -(b: Complex[T])(implicit r: Rng[T]): Complex[T] =
+  def -(b: Complex[T])(implicit r: CRing[T]): Complex[T] =
     new Complex(real - b.real, imag - b.imag)
 
-  def *(b: Complex[T])(implicit r: Rng[T]): Complex[T] =
+  def *(b: Complex[T])(implicit r: CRing[T]): Complex[T] =
     new Complex(real * b.real - imag * b.imag, imag * b.real + real * b.imag)
 
   def /(b: Complex[T])(implicit f: Field[T], s: Signed[T]): Complex[T] = {
@@ -348,252 +356,25 @@ final case class Complex[@sp(Float, Double) T](real: T, imag: T)
 }
 
 
-object FloatComplex {
-  import FastComplex.{encode}
 
-  final def apply(real: Float, imag: Float): FloatComplex =
-    new FloatComplex(encode(real, imag))
-
-  final def apply(real: Double, imag: Double): FloatComplex =
-    new FloatComplex(encode(real.toFloat, imag.toFloat))
-
-  def polar(magnitude: Float, angle: Float): FloatComplex =
-    new FloatComplex(FastComplex.polar(magnitude, angle))
-
-  final val i: FloatComplex = new FloatComplex(4575657221408423936L)
-  final val one: FloatComplex = new FloatComplex(1065353216L)
-  final val zero: FloatComplex = new FloatComplex(0L)
-}
-
-/**
- * Value class which encodes two floating point values in a Long.
- *
- * We get (basically) unboxed complex numbers using this hack.
- * The underlying implementation lives in the FastComplex object.
- */
-class FloatComplex(val u: Long) extends AnyVal {
-  override final def toString: String = "(%s+%si)" format (real, imag)
-
-  final def real: Float = FastComplex.real(u)
-  final def imag: Float = FastComplex.imag(u)
-  final def repr: String = "FloatComplex(%s, %s)" format(real, imag)
-  final def abs: Float = FastComplex.abs(u)
-  final def angle: Float = FastComplex.angle(u)
-  final def conjugate: FloatComplex = new FloatComplex(FastComplex.conjugate(u))
-  final def isWhole: Boolean = FastComplex.isWhole(u)
-  final def signum: Int = FastComplex.signum(u)
-  final def complexSignum: FloatComplex = new FloatComplex(FastComplex.complexSignum(u))
-  final def negate: FloatComplex = new FloatComplex(FastComplex.negate(u))
-
-  final def +(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.add(u, b.u))
-  final def -(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.subtract(u, b.u))
-  final def *(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.multiply(u, b.u))
-  final def /(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.divide(u, b.u))
-
-    /* TODO: does it make sense? Should match the behavior on Gaussian integers.
-  final def /~(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.quot(u, b.u))
-  final def %(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.mod(u, b.u))
-
-  final def /%(b: FloatComplex): (FloatComplex, FloatComplex) = FastComplex.quotmod(u, b.u) match {
-    case (q, m) => (new FloatComplex(q), new FloatComplex(m))
-  }
-     */
-
-  final def pow(b: FloatComplex): FloatComplex = new FloatComplex(FastComplex.pow(u, b.u))
-  final def **(b: FloatComplex): FloatComplex = pow(b)
-
-  final def pow(b: Int): FloatComplex = new FloatComplex(FastComplex.pow(u, FastComplex(b.toFloat, 0.0F)))
-  final def **(b: Int): FloatComplex = pow(b)
-}
-
-
-/**
- * FastComplex is an ugly, beautiful hack.
- *
- * The basic idea is to encode two 32-bit Floats into a single 64-bit Long.
- * The lower-32 bits are the "real" Float and the upper-32 are the "imaginary"
- * Float.
- *
- * Since we're overloading the meaning of Long, all the operations have to be
- * defined on the FastComplex object, meaning the syntax for using this is a
- * bit ugly. To add to the ugly beauty of the whole thing I could imagine
- * defining implicit operators on Long like +@, -@, *@, /@, etc.
- *
- * You might wonder why it's even worth doing this. The answer is that when
- * you need to allocate an array of e.g. 10-20 million complex numbers, the GC
- * overhead of using *any* object is HUGE. Since we can't build our own
- * "pass-by-value" types on the JVM we are stuck doing an encoding like this.
- *
- * Here are some profiling numbers for summing an array of complex numbers,
- * timed against a concrete case class implementation using Float (in ms):
- *
- *  size | encoded |  class
- *    1M |     5.1 |    5.8
- *    5M |    28.5 |   91.7
- *   10M |    67.7 |  828.1
- *   20M |   228.0 | 2687.0
- *
- * Not bad, eh?
- */
-object FastComplex {
-  import java.lang.Math.{atan2, cos, sin}
-
-  // note the superstitious use of @inline and final everywhere
-
-  final def apply(real: Float, imag: Float): Long = encode(real, imag)
-  final def apply(real: Double, imag: Double): Long = encode(real.toFloat, imag.toFloat)
-
-  // encode a float as some bits
-  @inline final def bits(n: Float): Int = java.lang.Float.floatToIntBits(n)
-
-  // decode some bits into a float
-  @inline final def bits(n: Int): Float = java.lang.Float.intBitsToFloat(n)
-
-  // get the real part of the complex number
-  @inline final def real(d: Long): Float = bits((d & 0xffffffff).toInt)
-
-  // get the imaginary part of the complex number
-  @inline final def imag(d: Long): Float = bits((d >>> 32).toInt)
-
-  // define some handy constants
-  final val i: Long = encode(0.0F, 1.0F)
-  final val one: Long = encode(1.0F, 0.0F)
-  final val zero: Long = encode(0.0F, 0.0F)
-
-  // encode two floats representing a complex number
-  @inline final def encode(real: Float, imag: Float): Long =
-    (bits(real) & 0xffffffffL) | ((bits(imag) & 0xffffffffL) << 32)
-
-  // encode two floats representing a complex number in polar form
-  @inline final def polar(magnitude: Float, angle: Float): Long =
-    encode(magnitude * cos(angle).toFloat, magnitude * sin(angle).toFloat)
-
-  // decode should be avoided in fast code because it allocates a Tuple2.
-  final def decode(d: Long): (Float, Float) = (real(d), imag(d))
-
-  // produces a string representation of the Long/(Float,Float)
-  final def toRepr(d: Long): String = "FastComplex(%s -> %s)" format(d, decode(d))
-
-  // get the magnitude/absolute value
-  final def abs(d: Long): Float = {
-    val re = real(d)
-    val im = imag(d)
-    java.lang.Math.sqrt(re * re + im * im).toFloat
-  }
-
-  // get the angle/argument
-  final def angle(d: Long): Float = atan2(imag(d), real(d)).toFloat
-
-  // get the complex conjugate
-  final def conjugate(d: Long): Long = encode(real(d), -imag(d))
-
-  // see if the complex number is a whole value
-  final def isWhole(d: Long): Boolean = real(d) % 1.0F == 0.0F && imag(d) % 1.0F == 0.0F
-
-  // get the sign of the complex number
-  final def signum(d: Long): Int = real(d) compare 0.0F
-
-  // get the complex sign of the complex number
-  final def complexSignum(d: Long): Long = {
-    val m = abs(d)
-    if (m == 0.0F) zero else divide(d, encode(m, 0.0F))
-  }
-
-  // negation
-  final def negate(a: Long): Long = encode(-real(a), -imag(a))
-
-  // addition
-  final def add(a: Long, b: Long): Long = encode(real(a) + real(b), imag(a) + imag(b))
-
-  // subtraction
-  final def subtract(a: Long, b: Long): Long = encode(real(a) - real(b), imag(a) - imag(b))
-
-  // multiplication
-  final def multiply(a: Long, b: Long): Long = {
-    val re_a = real(a)
-    val im_a = imag(a)
-    val re_b = real(b)
-    val im_b = imag(b)
-    encode(re_a * re_b - im_a * im_b, im_a * re_b + re_a * im_b)
-  }
-
-  // division
-  final def divide(a: Long, b: Long): Long = {
-    val re_a = real(a)
-    val im_a = imag(a)
-    val re_b = real(b)
-    val im_b = imag(b)
-
-    val abs_re_b = Math.abs(re_b)
-    val abs_im_b = Math.abs(im_b)
-
-    if (abs_re_b >= abs_im_b) {
-      if (abs_re_b == 0.0F) throw new ArithmeticException("/0")
-      val ratio = im_b / re_b
-      val denom = re_b + im_b * ratio
-      encode((re_a + im_a * ratio) / denom, (im_a - re_a * ratio) / denom)
-
-    } else {
-      if (abs_im_b == 0.0F) throw new ArithmeticException("/0")
-      val ratio = re_b / im_b
-      val denom = re_b * ratio + im_b
-      encode((re_a * ratio + im_a) / denom, (im_a * ratio - re_a) / denom)
-    }
-  }
-
-  /* TODO: does it make sense? Should match the behvaior on Gaussian integers
-   final def quot(a: Long, b: Long): Long =
-    encode(Math.floor(real(divide(a, b))).toFloat, 0.0F)
-
-  final def mod(a: Long, b: Long): Long = subtract(a, multiply(b, quot(a, b)))
-
-  final def quotmod(a: Long, b: Long): (Long, Long) = {
-    val q = quot(a, b)
-    (q, subtract(a, multiply(b, quot(a, b))))
-  }
-   */
-
-  // exponentiation
-  final def pow(a: Long, b: Long): Long = if (b == zero) {
-    encode(1.0F, 0.0F)
-
-  } else if (a == zero) {
-    if (imag(b) != 0.0F || real(b) < 0.0F)
-      throw new Exception("raising 0 to negative/complex power")
-    zero
-
-  } else if (imag(b) != 0.0F) {
-    val im_b = imag(b)
-    val re_b = real(b)
-    val len = (Math.pow(abs(a), re_b) / exp((angle(a) * im_b))).toFloat
-    val phase = (angle(a) * re_b + log(abs(a)) * im_b).toFloat
-    polar(len, phase)
-
-  } else {
-    val len = Math.pow(abs(a), real(b)).toFloat
-    val phase = (angle(a) * real(b)).toFloat
-    polar(len, phase)
-  }
-}
 
 trait ComplexInstances0 {
-  implicit def ComplexRing[A: Ring: Signed]: Ring[Complex[A]] = new ComplexIsRingImpl[A]
+  implicit def ComplexCRing[A: CRing: Signed]: CRing[Complex[A]] with Conjugation[Complex[A]] = new ComplexIsCRingImpl[A]
 }
 
 trait ComplexInstances1 extends ComplexInstances0 {
-  implicit def ComplexField[A: Field: Signed]: Field[Complex[A]] = new ComplexIsFieldImpl[A]
+  implicit def ComplexField[A: Field: Signed]: Field[Complex[A]] with Conjugation[Complex[A]] = new ComplexIsFieldImpl[A]
 }
 
 trait ComplexInstances extends ComplexInstances1 {
-  implicit def ComplexAlgebra[@sp(Float, Double) A: Fractional: Trig: Signed]: ComplexAlgebra[A] =
+  implicit def ComplexAlgebra[@sp(Float, Double) A: Fractional: Trig: Signed]: ComplexAlgebra[A] with Conjugation[Complex[A]] =
     new ComplexAlgebra[A]
 
   implicit def ComplexEq[A: Eq]: Eq[Complex[A]] = new ComplexEq[A]
 }
 
-private[math] trait ComplexIsRing[@sp(Float, Double) A] extends Ring[Complex[A]] {
-  implicit def algebra: Ring[A]
-  implicit def signed: Signed[A]
+private[math] trait ComplexIsCRing[@sp(Float, Double) A] extends CRing[Complex[A]] with Conjugation[Complex[A]] {
+  implicit def algebra: CRing[A]
 
   override def minus(a: Complex[A], b: Complex[A]): Complex[A] = a - b
   def negate(a: Complex[A]): Complex[A] = -a
@@ -602,12 +383,15 @@ private[math] trait ComplexIsRing[@sp(Float, Double) A] extends Ring[Complex[A]]
   override def times(a: Complex[A], b: Complex[A]): Complex[A] = a * b
   def zero: Complex[A] = Complex.zero
 
+  def conjugate(a: Complex[A]): Complex[A] = a.conjugate
+
   override def fromInt(n: Int): Complex[A] = Complex.fromInt[A](n)
 }
 
-private[math] trait ComplexIsField[@sp(Float,Double) A] extends ComplexIsRing[A] with Field.WithDefaultGCD[Complex[A]] {
+private[math] trait ComplexIsField[@sp(Float,Double) A] extends ComplexIsCRing[A] with Field.WithDefaultGCD[Complex[A]] {
 
   implicit def algebra: Field[A]
+  implicit def signed: Signed[A]
 
   override def fromDouble(n: Double): Complex[A] = Complex(algebra.fromDouble(n))
   def div(a: Complex[A], b: Complex[A]): Complex[A] = a / b
@@ -663,8 +447,7 @@ private[math] class ComplexEq[A: Eq] extends Eq[Complex[A]] with Serializable {
 }
 
 @SerialVersionUID(1L)
-private[math] final class ComplexIsRingImpl[@sp(Float,Double) A](implicit
-    val algebra: Ring[A], val signed: Signed[A]) extends ComplexIsRing[A] with Serializable
+private[math] final class ComplexIsCRingImpl[@sp(Float,Double) A](implicit val algebra: CRing[A]) extends ComplexIsCRing[A] with Serializable
 
 @SerialVersionUID(1L)
 private[math] final class ComplexIsFieldImpl[@sp(Float,Double) A](implicit
