+++ laws/src/main/scala/spire/laws/shadows/Shadow.scala
@@ -31,6 +31,13 @@ abstract class ShadowInstances0 {
       def A = implicitly
       def S = implicitly
     }
+
+  implicit def conjugationInstance[A:Conjugation, S:Conjugation](implicit ev: Shadowing[A, S]): Conjugation[Shadow[A, S]] =
+    new ShadowConjugation[A, S] {
+      def A: Conjugation[A] = implicitly
+      def S: Conjugation[S] = implicitly
+      val shadowing: Shadowing[A, S] = ev
+    }
 }
