+++ tests/src/test/scala/spire/SyntaxTest.scala
@@ -53,6 +53,7 @@ class SyntaxTest extends SpireTests with Checkers with BaseSyntaxTest {
   test("Order syntax")(check(forAll { (a: Int, b: Int) => testOrderSyntax(a, b) }))
   test("Signed syntax")(check(forAll { (a: Int) => testSignedSyntax(a) }))
   test("TruncatedDivision syntax")(check(forAll { (a: Int, b: NonZero[Int]) => testTruncatedDivisionSyntax(a, b.x) }))
+  test("Conjugation syntax")(check(forAll { (a: Int) => testConjugationSyntax(a) }))
   test("IsReal syntax")(check(forAll { (a: Double) => testIsRealSyntax(a) }))
   test("Semigroup syntax")(check(forAll { (a: String, b: String) => testSemigroupSyntax(a, b) }))
   test("Monoid syntax")(check(forAll { (a: String, b: String) => testMonoidSyntax(a, b) }))
@@ -150,6 +151,12 @@ trait BaseSyntaxTest {
     ((a fquot b) === TruncatedDivision[A].fquot(a, b)) &&
     ((a fmod b) === TruncatedDivision[A].fmod(a, b)) &&
     ((a fquotmod b) === TruncatedDivision[A].fquotmod(a, b))
+  }
+
+  def testConjugationSyntax[A: Conjugation: Eq](a: A) = {
+    import spire.syntax.conjugation._
+    import spire.syntax.eq._
+    a.conjugate === Conjugation[A].conjugate(a)
   }
 
   def testIsRealSyntax[A: IsReal](a: A) = {
